requirements.txt
asgiref==3.7.2
Django==4.2.7
sqlparse==0.4.4
tzdata==2023.3
pytz==2023.3.post1
djangorestframework==3.14.0
requests==2.31.0
beautifulsoup4==4.12.2